api/requirements.txt
APScheduler==3.9.1
attrs==21.4.0
black==22.1.0
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
flake8==4.0.1
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.1
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.7
pytest==7.1.1
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
sentry-sdk==1.5.8
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.32
tomli==2.0.1
typing-extensions==4.1.1
tzdata==2022.1
tzlocal==4.1
urllib3==1.26.9
Werkzeug==2.0.3
